% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{combine_gbif_records}
\alias{combine_gbif_records}
\title{This function combines all occurrence files from GBIF into one table.
There are slight differences between ALA and GBIF, so separate functions are useful.
It assumes that all the files come from the previous GBIF downloads function.
Although you can download all the records for in one go, this is better for
Doing small runs of species, or where you want to re-run them constantly}
\usage{
combine_gbif_records(
  species_list,
  records_path,
  records_extension,
  record_type,
  keep_cols,
  world_raster
)
}
\arguments{
\item{species_list}{List of species already downloaded}

\item{records_path}{File path for downloaded species}

\item{records_extension}{Which R file type? RDS or RDA}

\item{record_type}{Adds a column to the data frame for the data source, EG ALA}

\item{keep_cols}{The columns we want to keep - a character list created by you}

\item{world_raster}{An Raster file of the enviro conditions used (assumed to be global)}
}
\description{
This function combines all occurrence files from GBIF into one table.
There are slight differences between ALA and GBIF, so separate functions are useful.
It assumes that all the files come from the previous GBIF downloads function.
Although you can download all the records for in one go, this is better for
Doing small runs of species, or where you want to re-run them constantly
}
