% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAXENT_FUNCTIONS.R
\name{compile_sdm_results}
\alias{compile_sdm_results}
\title{This function extracts the SDM results from the folders.
It assumes that the input folders are those returned by the 'fit_maxent_targ_bg_back_sel' function}
\usage{
compile_sdm_results(species_list, results_dir, save_data, data_path, save_run)
}
\arguments{
\item{species_list}{Character string - the species to run maxent models for}

\item{results_dir}{Character string - The file path used for saving the maxent output}

\item{save_data}{Logical or character - do you want to save the data frame?}

\item{data_path}{Character string - The file path used for saving the data frame}

\item{save_run}{Character string - run name to append to the data frame, useful for multiple runs.}
}
\value{
Data.frame of maxent results
}
\description{
This function extracts the SDM results from the folders.
It assumes that the input folders are those returned by the 'fit_maxent_targ_bg_back_sel' function
}
