% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{coord_clean_records}
\alias{coord_clean_records}
\title{This function takes a data frame of all species records, and flag records as institutional or spatial outliers.
It uses the CoordinateCleaner package https://cran.r-project.org/web/packages/CoordinateCleaner/index.html.
It assumes that the records data.frame is that returned by the combine_records_extract function}
\usage{
coord_clean_records(
  records,
  capitals,
  centroids,
  save_run,
  data_path,
  save_data
)
}
\arguments{
\item{records}{Data.frame. DF of all species records returned by the combine_records_extract function}

\item{capitals}{Numeric. Remove records within an xkm radius of capital cites (see ?clean_coordinates)}

\item{centroids}{Numeric. Remove records within an xkm radius around country centroids (see ?clean_coordinates)}

\item{save_run}{Character string - run name to append to the data frame, useful for multiple runs.}

\item{data_path}{Character string - The file path used for saving the data frame}

\item{save_data}{Logical or character - do you want to save the data frame?}
}
\value{
Data.frame of all urban records, with global enviro conditions for each record location (i.e. lat/lon)
}
\description{
This function takes a data frame of all species records, and flag records as institutional or spatial outliers.
It uses the CoordinateCleaner package https://cran.r-project.org/web/packages/CoordinateCleaner/index.html.
It assumes that the records data.frame is that returned by the combine_records_extract function
}
