% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{download_ALA_all_species}
\alias{download_ALA_all_species}
\title{Download species occurrence files from the Atlas of Living Australia (ALA)}
\usage{
download_ALA_all_species(
  species_list,
  your_email,
  download_path,
  download_limit
)
}
\arguments{
\item{species_list}{Character vector - List of species binomials to download}

\item{download_path}{Character string - File path for species downloads}

\item{download_limit}{Numeric - How many records can be downloaded at one time? Set by server}
}
\description{
This function downloads species occurrence files from ALA (https://www.ala.org.au/).
It assumes that the species list supplied is taxonomically correct.
It downloads the species without returning anything
}
