% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAXENT_FUNCTIONS.R
\name{fit_maxent_targ_bg_back_sel}
\alias{fit_maxent_targ_bg_back_sel}
\title{This function takes a data frame of all species records,
and runs a specialised maxent analysis for each species.
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the input df is that returned by the prepare_sdm_table function}
\usage{
fit_maxent_targ_bg_back_sel(
  occ,
  bg,
  sdm_predictors,
  name,
  outdir,
  bsdir,
  cor_thr,
  pct_thr,
  k_thr,
  backwards_sel,
  template_raster,
  min_n,
  max_bg_size,
  background_buffer_width,
  Koppen_raster,
  shapefiles,
  features,
  replicates,
  responsecurves,
  country_shp
)
}
\arguments{
\item{occ}{SpatialPointsDataFrame - Spdf of all species records returned by the 'prepare_sdm_table' function}

\item{bg}{SpatialPointsDataFrame - Spdf of of candidate background points}

\item{sdm_predictors}{Character string - Vector of enviro conditions that you want to include}

\item{outdir}{Character string - The file path used for saving the maxent output}

\item{bsdir}{Character string - The file path used for saving the backwards selection maxent output}

\item{cor_thr}{Numeric - The max allowable pairwise correlation between predictor variables}

\item{pct_thr}{Numeric - The min allowable percent variable contribution}

\item{k_thr}{Numeric - The min number of variables to be kept in the model}

\item{backwards_sel}{Logical - Run backwards selection using the maxent models (T/F)?}

\item{template_raster}{RasterLayer -  Empty raster with analysis extent (global), res (1km) and projection (Mollweide, EPSG 54009)}

\item{min_n}{Numeric - The min number of records for running maxent}

\item{max_bg_size}{Numeric - The max number of background points to keep}

\item{background_buffer_width}{Numeric - The max distance (km) from occ points that BG points should be selected?}

\item{shapefiles}{Logical - Save shapefiles of the occ and bg data (T/F)?}

\item{features}{Character string - Which features should be used? (e.g. linear, product, quadratic 'lpq')}

\item{replicates}{Numeric - The number of replicates to use}

\item{responsecurves}{Logical - Save response curves of the maxent models (T/F)?}

\item{country_shp}{.Rds object - SpatialPolygonsDataFrame of Australia for mapping maxent points}

\item{rep_args}{RasterLayer of global koppen zones, in Mollweide54009 projection}

\item{full_args}{Dataframe of global koppen zones, with columns : GRIDCODE, Koppen}
}
\description{
This function takes a data frame of all species records,
and runs a specialised maxent analysis for each species.
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the input df is that returned by the prepare_sdm_table function
}
