% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAXENT_FUNCTIONS.R
\name{local_simplify}
\alias{local_simplify}
\title{This is a local version of rmaxent::simplify (https://github.com/johnbaums/rmaxent)
Given a candidate set of predictor variables, this function identifies
a subset that meets specified multicollinearity criteria. Subsequently,
backward stepwise variable selection is used to iteratively drop the variable
that contributes least to the model, until the contribution of each variable
meets a specified minimum, or until a predetermined minimum number of predictors remains.
It assumes that the input df is that returned by the fit_maxent_targ_bg_back_sel function}
\usage{
local_simplify(
  occ,
  bg,
  path,
  species_column = "species",
  response_curves = TRUE,
  logistic_format = TRUE,
  type = "PI",
  cor_thr,
  pct_thr,
  k_thr,
  features = "lpq",
  replicates = 1,
  quiet = TRUE
)
}
\arguments{
\item{occ}{SpatialPointsDataFrame - Spdf of all species records returned by the 'prepare_sdm_table' function}

\item{bg}{SpatialPointsDataFrame - Spdf of of candidate background points}

\item{path}{Character string - Vector of enviro conditions that you want to include}

\item{species_column}{Character string - Vector of enviro conditions that you want to include}

\item{logistic_format}{Logical value indicating whether maxentResults.csv should report logistic value thresholds}

\item{type}{The variable contribution metric to use when dropping variables}

\item{cor_thr}{Numeric - The max allowable pairwise correlation between predictor variables}

\item{pct_thr}{Numeric - The min allowable percent variable contribution}

\item{k_thr}{Numeric - The min number of variables to be kept in the model}

\item{features}{Character string - Which features should be used? (e.g. linear, product, quadratic 'lpq')}

\item{replicates}{Numeric - The number of replicates to use}

\item{responsecurves}{Logical - Save response curves of the maxent models (T/F)?}
}
\description{
This is a local version of rmaxent::simplify (https://github.com/johnbaums/rmaxent)
Given a candidate set of predictor variables, this function identifies
a subset that meets specified multicollinearity criteria. Subsequently,
backward stepwise variable selection is used to iteratively drop the variable
that contributes least to the model, until the contribution of each variable
meets a specified minimum, or until a predetermined minimum number of predictors remains.
It assumes that the input df is that returned by the fit_maxent_targ_bg_back_sel function
}
