% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{prepare_sdm_table}
\alias{prepare_sdm_table}
\title{This function takes a data frame of all species records,
And prepares a table in the 'species with data' (swd) format for modelling uses the Maxent algorithm.
It assumes that the input df is that returned by the coord_clean_records function}
\usage{
prepare_sdm_table(
  coord_df,
  species_list,
  BG_points,
  sdm_table_vars,
  save_run,
  save_shp,
  read_background,
  save_data,
  data_path
)
}
\arguments{
\item{coord_df}{Data.frame. DF of all species records returned by the coord_clean_records function}

\item{species_list}{Character string - the species analysed}

\item{BG_points}{Logical - Do we want to include a dataframe of background points? Otherwise, BG points taken from species not modelled}

\item{sdm_table_vars}{Character string - The enviro vars to be included in species models}

\item{save_run}{Character string - append a run name to the output (e.g. 'bat_species')}

\item{save_shp}{Logical - Save a shapefile of the SDM table (T/F)?}

\item{read_background}{Logical - Read in an additional dataframe of background points (T/F)?}

\item{save_data}{Logical - do you want to save the data frame?}

\item{data_path}{Character string - The file path used for saving the data frame}
}
\value{
Data.frame of species records, with spatial outlier T/F flag for each record
}
\description{
This function takes a data frame of all species records,
And prepares a table in the 'species with data' (swd) format for modelling uses the Maxent algorithm.
It assumes that the input df is that returned by the coord_clean_records function
}
