% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{project_maxent_grids_mess}
\alias{project_maxent_grids_mess}
\title{This function takes the maxent models created by the 'fit_maxent_targ_bg_back_sel' function,
and projects the model across geographic space - currently just for Australia.
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the maxent models were generated by the 'fit_maxent_targ_bg_back_sel'}
\usage{
project_maxent_grids_mess(
  country_shp,
  world_shp,
  country_prj,
  world_prj,
  local_prj,
  scen_list,
  species_list,
  maxent_path,
  climate_path,
  grid_names,
  time_slice,
  current_grids,
  create_mess,
  nclust,
  OSGeo_path
)
}
\arguments{
\item{country_shp}{SpatialPolygonsDataFrame - Spdf of the country for mapping maxent results (e.g. Australia)}

\item{world_shp}{SpatialPolygonsDataFrame - Spdf of the world for mapping maxent results}

\item{country_prj}{CRS object  - Local projection for mapping maxent results}

\item{world_prj}{CRS object  - Global projection for mapping maxent results}

\item{local_prj}{CRS object  - Local projection for mapping maxent results}

\item{scen_list}{Character string - The list of global circulation models to create predictions for}

\item{species_list}{Character string - The species to run maxent predictions for}

\item{maxent_path}{Character string - The file path containin the existing maxent models}

\item{climate_path}{Character string - The file path where the climate data is saved}

\item{grid_names}{Character string - Vector of enviro conditions that you want to include}

\item{time_slice}{Character string - The time period to create predictions for (e.g. '2050', or '2070')}

\item{current_grids}{Character string - Vector of current enviro conditions that you want to include}

\item{create_mess}{Logical - Create mess maps of the preditions (T/F)?}

\item{nclust}{Numeric - How many clusters to use for parallel processing (e.g. 1)}

\item{OSGeo_path}{Character string - file path of .bat file for converting rasters to polygons (
you need to download the OSGeo4W64 setup, see https://www.osgeo.org/)}
}
\description{
This function takes the maxent models created by the 'fit_maxent_targ_bg_back_sel' function,
and projects the model across geographic space - currently just for Australia.
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the maxent models were generated by the 'fit_maxent_targ_bg_back_sel'
}
