% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{sdm_area_cell_count}
\alias{sdm_area_cell_count}
\title{This  function uses the 10th% Logistic threshold for each species from the maxent
models to threhsold the rasters of habitat suitability (0-1) For each GCM. For each
species, summ the 6 GCMS to create a binary raster with cell values between 0-6.
These cell values represent the number of GCMs where that cell had a suitability
value above the threshold determined by maxent. We classify a cell has suitable
if it met the threshold in > 4 GCMs, and use this combined raster to compare
current and future suitability, measuring if the suitability of each cell is
changing over time, remaining stable or was never suitable
It assumes that the maxent predictions were generated by the 'project_maxent_grids_mess' function}
\usage{
sdm_area_cell_count(
  unit_shp,
  country_shp,
  world_shp,
  sort_var,
  agg_var,
  unit_vec,
  DIR_list,
  species_list,
  number_gcms,
  maxent_path,
  thresholds,
  time_slice,
  write_rasters
)
}
\arguments{
\item{unit_shp}{SpatialPolygonsDataFrame - Spdf of the spatial units used to aggregate the maxent results (e.g. Australia)}

\item{country_shp}{SpatialPolygonsDataFrame - Spdf of the country for aggregating maxent results (e.g. Australia)}

\item{world_shp}{SpatialPolygonsDataFrame - Spdf of the world for aggregating maxent results}

\item{sort_var}{Character string - The field name in the shapefile to use for sorting (e.g. Urban area names)}

\item{agg_var}{Character string - The field name in the shapefile to use for aggregating SDM results (e.g. Urban area codes)}

\item{species_list}{Character string - The species to run maxent predictions for}

\item{maxent_path}{Character string - The file path containin the existing maxent models}

\item{time_slice}{Character string - The time period to create predictions for (e.g. '2050', or '2070')}

\item{@param}{unit_vec    Character string - The field name in the shapefile to use for aggregating SDM results (e.g. Urban area codes)}

\item{scen_list}{Character string - The list of global circulation models to create predictions for}
}
\description{
This  function uses the 10th% Logistic threshold for each species from the maxent
models to threhsold the rasters of habitat suitability (0-1) For each GCM. For each
species, summ the 6 GCMS to create a binary raster with cell values between 0-6.
These cell values represent the number of GCMs where that cell had a suitability
value above the threshold determined by maxent. We classify a cell has suitable
if it met the threshold in > 4 GCMs, and use this combined raster to compare
current and future suitability, measuring if the suitability of each cell is
changing over time, remaining stable or was never suitable
It assumes that the maxent predictions were generated by the 'project_maxent_grids_mess' function
}
