% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{shapefile_vector_from_raster}
\alias{shapefile_vector_from_raster}
\title{This function takes a raster of a shapefile (for example the urban areas of Australia),
and creates a shapefile (i.e. a vector).
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the input df is that returned by the prepare_sdm_table function}
\usage{
shapefile_vector_from_raster(shp_file, prj, agg_var, temp_ras, targ_ras)
}
\arguments{
\item{shp_file}{SpatialPolygonsDataFrame - Spdf of spatial units used to aggregate the SDMs (e.g. urban areas of Australia)}

\item{prj}{CRS object - Local projection for mapping the shapefile (e.g. Australian Albers)}

\item{agg_var}{Character string - The field name in the shapefile to use for aggregating SDM results (e.g. Urban area codes)}

\item{temp_ras}{Raster - An existing raster with the same extent, resolution and projection as the maxent models (e.g. Australia)}

\item{targ_ras}{Raster - An existing raster of the shapefile with the same extent, resolution and projection as the maxent models (e.g. Australia)}

\item{sort_var}{Character string - The field name in the shapefile to use for sorting (e.g. Urban area names)}
}
\description{
This function takes a raster of a shapefile (for example the urban areas of Australia),
and creates a shapefile (i.e. a vector).
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the input df is that returned by the prepare_sdm_table function
}
