% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{urban_records_extract}
\alias{urban_records_extract}
\title{This function takes a data frame from Ubran sources (e.g. I-naturalist), and extracts enviro values.
It assumes that the urban dataframe has these columns : species, lat, lon, Country, INVENTORY, SOURCE}
\usage{
urban_records_extract(
  urban_df,
  species_list,
  thin_records,
  template_raster,
  world_raster,
  prj,
  biocl_vars,
  env_vars,
  worldclim_grids,
  save_data,
  data_path,
  save_run
)
}
\arguments{
\item{urban_df}{Data.frame of Urban records (only used if you have urban data, e.g. I-naturalist)}

\item{species_list}{Character string - List of species analysed, used to cut the dataframe down}

\item{thin_records}{Do you want to thin the records out? If so, it will be 1 record per 1km*1km grid cell}

\item{template_raster}{A global R Raster used to thin records to 1 record per 1km grid cell}

\item{world_raster}{An global R Raster of the enviro conditions used to extract values for all records}

\item{prj}{The projection system used. Currently, needs to be WGS84}

\item{biocl_vars}{The variables used - eg the standard bioclim names (https://www.worldclim.org/).}

\item{env_vars}{The actual anmes of the variables (e.g. bio1 = rainfall, etc.) Only needed for worldlcim}

\item{worldclim_grids}{Are you using worldclim stored as long intergers? If so, divide by 10.}

\item{save_data}{Do you want to save the data frame?}

\item{data_path}{The file path used for saving the data frame}

\item{save_run}{Character string - run name to append to the data frame (e.g. bat species, etc.). Useful for multiple runs.}
}
\value{
Data frame of all urban records, with global enviro conditions for each record location (i.e. lat/lon)
}
\description{
This function takes a data frame from Ubran sources (e.g. I-naturalist), and extracts enviro values.
It assumes that the urban dataframe has these columns : species, lat, lon, Country, INVENTORY, SOURCE
}
